msc: Separate the BSC and MSC link completly...
Make the msc_conn responsible for creating the link to
the core network and reopening it, make the BTS code just
call the msc methods and the MSC will throw away data in
case it can not be forwarded.

This avoids a problem that we start a reconnect timer
while we have a connection in progress and then add the
same file descriptor twice. This is mostly a speculative
fix to the problem.