mtp: Rename link_data to mtp_link and move out the transport specific things
Rename link_data to mtp_link and move it into the mtp_data header
file, also remove the union to ease creating more of the subtypes.

This is done in preparation to the linkset knowing more about the
link (e.g. having a link test per link instead of per link).