link_sets: Remove direct function calls for data/up/down from MTP3
We want to support different link_sets types. To do this we will
need to reduce the coupling of mtp_link_set with the application.
This is the first step to remove the notifications. This comes with
a performance cost but it has not been measured.

In the long run the conditions for if (set->on_bla) should be
removed as this indicates a wrong configuration of the linkset.