m3ua/sccp: Allow to skip T18/T20 TFP/TRA handling completely
For M3UA we already skip the link-test handling and there is no
point in waiting for management messages to collect routing info.
We already know where to send things. At first I tried to overload
the skip handling but it is better to just set the timer to 0 and
then move on.